move Button component out of lab
