+++ .changeset/ninety-cherries-promise.md
@@ -1,0 +1,5 @@
+---
+'@utilitywarehouse/web-ui': major
+---
+
+Replace Button component
